python/voidvoid/client/requirements.txt
websocket-client==0.48.0
